Added support for ubuntu 24.04 and debian 12.

+++ rootfs/etc/cont-init.d/10-pkgs-mirror.sh
@@ -22,17 +22,28 @@ case "$ID" in
         ;;
     debian)
         echo "setting packages mirror to '$PACKAGES_MIRROR'..."
-        cp -a /defaults/sources.list /etc/apt/sources.list
-        sed-patch "s|^deb http://deb.debian.org/debian |deb $PACKAGES_MIRROR |g" /etc/apt/sources.list
+        if [ -f /defaults/debian.sources ]; then
+            cp -a /defaults/debian.sources /etc/apt/sources.list.d/debian.sources
+            sed-patch "s|^URIs: http://deb.debian.org/debian\$|URIs: $PACKAGES_MIRROR|g" /etc/apt/sources.list.d/debian.sources
+        else
+            cp -a /defaults/sources.list /etc/apt/sources.list
+            sed-patch "s|^deb http://deb.debian.org/debian |deb $PACKAGES_MIRROR |g" /etc/apt/sources.list
+        fi
         ;;
     ubuntu)
         echo "setting packages mirror to '$PACKAGES_MIRROR'..."
-        cp -a /defaults/sources.list /etc/apt/sources.list
-        if grep -q "http://ports.ubuntu.com/ubuntu-ports/" /etc/apt/sources.list
-        then
-            sed-patch "s|^deb http://ports.ubuntu.com/ubuntu-ports/ |deb $PACKAGES_MIRROR |g" /etc/apt/sources.list
+        if [ -f /defaults/ubuntu.sources ]; then
+            cp -a /defaults/ubuntu.sources /etc/apt/sources.list.d/ubuntu.sources
+            sed-patch "s|^URIs: http://archive.ubuntu.com/ubuntu/\$|URIs: $PACKAGES_MIRROR|g" /etc/apt/sources.list.d/ubuntu.sources
         else
-            sed-patch "s|^deb http://archive.ubuntu.com/ubuntu/ |deb $PACKAGES_MIRROR |g" /etc/apt/sources.list
+            cp -a /defaults/sources.list /etc/apt/sources.list
+            if grep -q "http://ports.ubuntu.com/ubuntu-ports/" /etc/apt/sources.list
+            then
+                # For archs other than i386/i686/x86_64.
+                sed-patch "s|^deb http://ports.ubuntu.com/ubuntu-ports/ |deb $PACKAGES_MIRROR |g" /etc/apt/sources.list
+            else
+                sed-patch "s|^deb http://archive.ubuntu.com/ubuntu/ |deb $PACKAGES_MIRROR |g" /etc/apt/sources.list
+            fi
         fi
         ;;
     *)
